arch: introduce architecture abstraction and move x86_64 code to arch/x86_64.c
