+++ client.c
@@ -2,6 +2,7 @@
 #include "xutil.h"
 #include "mem.h"
 #include "client.h"
+#include "arch.h"
 
 guest_client_t *guest_client = NULL;
 
@@ -18,11 +19,8 @@ int readmem(uint64_t addr, int memtype, void *buffer, long size)
 
     switch (memtype) {
         case KVADDR:
-            if (addr >= __START_KERNEL_map) {
-                paddr = ((addr) - (ulong)__START_KERNEL_map + machdep->machspec->phys_base);
-            } else {
-                paddr = ((addr) - PAGE_OFFSET);
-            }
+            if (arch_kvtop(addr, &paddr) < 0)
+                return -1;
             break;
         case PHYSADDR:
             paddr = addr;
@@ -39,6 +37,7 @@ int guest_client_new(char *ac, guest_access_t ty)
 
     guest_client_t *c = xcalloc(1, sizeof(guest_client_t));
     c->ty = ty;
+
     switch(c->ty) {
         case GUEST_NAME:
             if (libvirt_client_init(ac))
