+++ meson.build
@@ -6,9 +6,20 @@ cflags = ['-Wall', '-Wextra', '-O2']
 # Linker options
 ldflags = ['-ldl']
 
+# Architecture-specific source
+host_arch = host_machine.cpu_family()
+if host_arch == 'x86_64'
+  arch_src = 'arch/x86_64.c'
+elif host_arch == 'aarch64'
+  arch_src = 'arch/aarch64.c'
+else
+  error('unsupported architecture: ' + host_arch)
+endif
+
 # Sources
 sources = [
   'main.c',
+  arch_src,
   'log.c',
   'version.c',
   'kernel.c',
@@ -26,6 +37,6 @@ sources = [
 # Build executable
 executable('kvm-dmesg',
   sources,
-  c_args       : cflags,
-  link_args    : ldflags
+  c_args    : cflags,
+  link_args : ldflags
 )
